Allow anonymous users to get condor job status
